Bookmanagement CRUD

+++ BookManagement/frmBookManagement.cs
@@ -7,8 +7,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using AutoMapper;
 using BookObject.Models;
 using DataAccess.Repository;
+using BookObject.DTO;
+using BookObject.Mapper;
+using System.DirectoryServices;
+
 namespace BookManagement_TranQuocBao
 {
     public partial class frmBookManagement : Form
@@ -17,14 +22,45 @@ namespace BookManagement_TranQuocBao
         {
             InitializeComponent();
         }
+        public void EnableCreateAndUpdate()
+        {
+            btnCreate.Enabled = true;
+            btnUpdate.Enabled = true;
+            btnDelete.Enabled = false;
+        }
+
+        // Method to enable view-only functionality
+        public void EnableViewOnly()
+        {
+            btnCreate.Enabled = false;
+            btnUpdate.Enabled = false;
+            btnDelete.Enabled = false;
+        }
         private BookRepository bookRepository = new BookRepository();
         private BookCategoryRepository bookCategoryRepository = new BookCategoryRepository();
         private BindingSource bookSource;
 
-        public void LoadBook()
+        public void LoadBook(List<Book> books)
         {
             bookSource = new BindingSource();
-            var list = bookRepository.GetAllBook();
+            if (books != null && books.Count == 0)
+            {
+                dgvBook.DataSource = null;
+                return;
+            }
+
+            books = books == null ? bookRepository.GetAllBook() : books;
+            var config = new MapperConfiguration(cfg =>
+            {
+                BookConfig.createMap(cfg);
+            });
+
+            var mapperBookCon = config.CreateMapper();
+
+            var list = books.Select(
+                air => mapperBookCon
+                .Map<Book, BookDTO>(air)
+                );
             bookSource.DataSource = list;
 
             txtBookID.DataBindings.Clear();
@@ -46,7 +82,7 @@ namespace BookManagement_TranQuocBao
 
             dgvBook.DataSource = null;
             dgvBook.DataSource = bookSource;
-            dgvBook.Columns["BookCategory"].Visible = false;
+
             cboCategory.DataSource = bookCategoryRepository.GetBookCategories();
             cboCategory.DisplayMember = "BookGenreType";
             cboCategory.ValueMember = "BookCategoryId";
@@ -58,13 +94,20 @@ namespace BookManagement_TranQuocBao
 
         private void frmBookManagement_Load(object sender, EventArgs e)
         {
-            LoadBook();
+            LoadBook(null);
         }
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
-            bookRepository.delete(int.Parse(txtBookID.Text));
-            LoadBook();
+            DialogResult d;
+            d = MessageBox.Show("Do you ready want to delete", "Book Manager", MessageBoxButtons.OKCancel, MessageBoxIcon.Question, MessageBoxDefaultButton.Button1);
+            if (d == DialogResult.OK)
+            {
+                bookRepository.delete(int.Parse(txtBookID.Text));
+                LoadBook(null);
+            }
+
+
         }
 
         private void btnCreate_Click(object sender, EventArgs e)
@@ -73,14 +116,14 @@ namespace BookManagement_TranQuocBao
             {
                 Text = "Add Book",
                 InsertOrUpdate = false,
-                BookInfor = GetBook(),
+                
 
                 bookRepository = bookRepository,
                 bookCategoryRepository = bookCategoryRepository,
             };
             if (frmBook.ShowDialog() == DialogResult.OK)
             {
-                LoadBook();
+                LoadBook(null);
             }
         }
         public Book GetBook()
@@ -112,7 +155,7 @@ namespace BookManagement_TranQuocBao
             };
             if (frmBook.ShowDialog() == DialogResult.OK)
             {
-                LoadBook();
+                LoadBook(null);
             }
         }
 
@@ -137,7 +180,7 @@ namespace BookManagement_TranQuocBao
             else
             {
                 // If the search keyword is empty, reload all books
-                LoadBook();
+                LoadBook(null);
             }
         }
 
@@ -155,6 +198,16 @@ namespace BookManagement_TranQuocBao
             cboCategory.DisplayMember = "BookGenreType";
             cboCategory.ValueMember = "BookCategoryId";
             cboCategory.DataBindings.Add("SelectedValue", bookSource, "BookCategoryId");
+        }
+
+        private void btnLoad_Click(object sender, EventArgs e)
+        {
+            LoadBook(null);
+        }
+
+        private void btnLogout_Click(object sender, EventArgs e)
+        {
+            Application.Restart();
         }
     }
 }
